// src/types.ts
export interface Slide {
    image: string;
    title: string;
    subtitle: string;
}

export interface Service {
    icon: string;
    title: string;
    description: string;
}

export interface TeamMember {
    name: string;
    role: string;
    image: string;
    linkedin: string;
}

export interface Project {
    title: string;
    description: string;
    image: string;
    link: string;
}

export interface Testimonial {
    name: string;
    role: string;
    image: string;
    text: string;
}